Crate malachite

Modules

bools

Functions for working with bools.

chars

Functions for working with chars.

comparison

Macros and traits related to comparing values.

integer

Integer, a type representing integers with arbitrarily large absolute values.

iterators

Functions and adaptors for iterators.

named

The Named trait, for getting a type’s name.

natural

Natural, a type representing arbitrarily large non-negative integers.

nevers

Never, a type that cannot be instantiated.

num

Functions for working with primitive integers and floats.

options

Functions for working with Orderings.

orderings

Functions for working with Options.

rational_sequences

RationalSequence, a type representing a sequence that is finite or eventually repeating, just like the digits of a rational number.

rounding_modes

RoundingMode, an enum used to specify rounding behavior.

sets

Functions for working with HashSets and BTreeSets.

slices

Functions for working with slices.

strings

Functions for working with Strings.

tuples

Functions for working with tuples.

unions

Unions (sum types). These are essentially generic enums.

vecs

Functions for working with Vecs.

Macros

custom_tuples

Defines custom exhaustive tuple generators.

exhaustive_ordered_unique_tuples

Defines exhaustive ordered unique tuple generators.

exhaustive_tuples

Defines exhaustive tuple generators.

exhaustive_tuples_1_input

Defines exhaustive tuple generators that generate tuples from a single iterator.

exhaustive_unions

Defines exhaustive union generators.

exhaustive_unique_tuples

Defines lexicographic unique tuple generators.

exhaustive_vecs_fixed_length

Defines exhaustive fixed-length Vec generators.

impl_named

Automatically implements Named for a type.

lex_custom_tuples

Defines custom lexicographic tuple generators.

lex_ordered_unique_tuples

Defines lexicographic ordered unique tuple generators.

lex_tuples

Defines lexicographic tuple generators.

lex_unique_tuples

Defines lexicographic unique tuple generators.

lex_vecs_fixed_length

Defines lexicographic fixed-length Vec generators.

max

Computes the maximum of a list of expressions.

min

Computes the minimum of a list of expressions.

split_into_chunks

Splits an immutable slice into adjacent immutable chunks.

split_into_chunks_mut

Splits a mutable slice into adjacent mutable chunks.

union_struct

Defines unions.

Structs

Integer

An integer.

Natural

A natural (non-negative) integer.

Rational

A rational number.