Module malachite::chars::crement

Functions for incrementing and decrementing chars.

Functions

char_to_contiguous_range

Converts a char to a u32.

contiguous_range_to_char

Converts a u32 to a char; if all chars were arranged in ascending order, passing $u$ to this function would return the $u$th char.

decrement_char

Decrements this char, skipping over the surrogate code points.

increment_char

Increments this char, skipping over the surrogate code points.