Module malachite::integer::arithmetic

Traits for arithmetic.

Modules

abs

Absolute value of Integers, including implementations of UnsignedAbs.

add

Addition of Integers.

add_mul

Implementations of AddMul and AddMulAssign, traits for adding a number and the product of two other numbers.

binomial_coefficient

Implementations of BinomialCoefficient, a trait for computing the binomial coefficient of two numbers.

div

Division of Integers.

div_exact

Implementations of DivExact and DivExactAssign, traits for dividing two numbers when it’s known that the division is exact.

div_mod

Implementations of raits for simultaneously finding the quotient and remainder of two numbers, subject to various rounding rules.

div_round

Implementations of DivRound and DivExactAssign, traits for dividing two numbers according to a specified RoundingMode.

divisible_by

Implementations of DivisibleBy, a trait for determining whether one number is divisible by another.

divisible_by_power_of_2

Implementations of DivisibleByPowerOf2, a trait for determining whether a number is divisible by $2^k$.

eq_mod

Implementations of EqMod, a trait for determining whether one number is equal by another modulo a third.

eq_mod_power_of_2

Implementations of EqModPowerOf2, a trait for determining whether one number is equal to another modulo $2^k$.

extended_gcd

Implementations of ExtendedGcd, a trait for computing the extended GCD of two numbers.

kronecker_symbol

Implementations of LegendreSymbol, JacobiSymbol, and KroneckerSymbol, traits for computing the Legendre, Jacobi, and Kronecker symbols of two numbers.

mod_op

Implementations of traits for finding the remainder of two numbers, subject to various rounding rules.

mod_power_of_2

Implementations of traits for finding the remainder of a number divided by $2^k$, subject to various rounding rules.

mul

Multiplication of Integers.

neg

Negation of an Integer.

parity

Implementations of Parity, a trait for determining whether a number is even or odd.

pow

Implementations of Pow and PowAssign, traits for raising a number to a power.

power_of_2

Implementations of PowerOf2, a trait for computing a power of 2.

root

Implementations of traits for taking the $n$th root of a number.

round_to_multiple

Implementations of RoundToMultiple and RoundToMultipleAssign, traits for rounding a number to a multiple of another number.

round_to_multiple_of_power_of_2

Implementations of RoundToMultipleOfPowerOf2 and RoundToMultipleOfPowerOf2Assign, traits for rounding a number to a multiple of a power of 2.

shl

Left-shifting an Integer (multiplying it by a power of 2).

shl_round

Implementations of ShlRound and ShlRoundAssign, traits for multiplying a number by a power of 2 and rounding according to a specified RoundingMode.

shr

Right-shifting an Integer (dividing it by a power of 2).

shr_round

Implementations of ShrRound and ShrRoundAssign, traits for dividing a number by a power of 2 and rounding according to a specified RoundingMode.

sign

Implementations of Sign, a trait for determining the sign of a number.

sqrt

Implementations of traits for taking the square root of a number.

square

Implementations of Square and SquareAssign, traits for squaring a number.

sub

Subtraction of Integers.

sub_mul

Implementations of SubMul and SubMulAssign, traits for subtracting the product of two numbers from a number.