Module malachite::integer::conversion

Traits for converting to and from Integers, converting to and from strings, and extracting digits.

Modules

from_bool

An implementation of the From trait for converting a bool to an Integer.

from_natural

Functions for converting a Natural to an Integer, and an implementation of the From trait.

from_primitive_float

Implementations of traits for converting a primitive float to an Integer.

from_primitive_int

Implementations of traits for converting a primitive integer to an Integer.

from_twos_complement_limbs

Functions for constructing an Integer from two’s complement Limbs.

is_integer

An implementation of IsInteger, a trait for determining whether a number is an integer.

natural_from_integer

Implementations of traits for converting an Integer to a Natural.

primitive_float_from_integer

Implementations of traits for converting an Integer to a primitive float.

primitive_int_from_integer

Implementations of traits for converting an Integer to a primitive integer.

serde

Implementations of traits for serialization and deserialization using serde.

string

Implementations of traits for converting Integers to and from Strings.

to_twos_complement_limbs

Functions for extracting two’s complement Limbs from an Integer.