Trait malachite::num::arithmetic::traits::CheckedAddMul

pub trait CheckedAddMul<Y = Self, Z = Self> {
    type Output;

    // Required method
    fn checked_add_mul(self, y: Y, z: Z) -> Option<Self::Output>;
}

Adds a number and the product of two other numbers, returning None if the result is not representable.

Required Associated Types

type Output

Required Methods

fn checked_add_mul(self, y: Y, z: Z) -> Option<Self::Output>