Trait malachite::num::arithmetic::traits::DivMod

pub trait DivMod<RHS = Self> {
    type DivOutput;
    type ModOutput;

    // Required method
    fn div_mod(self, other: RHS) -> (Self::DivOutput, Self::ModOutput);
}

Divides two numbers, returning the quotient and remainder. The quotient is rounded towards negative infinity, and the remainder has the same sign as the divisor (second input).

The quotient and remainder satisfy $x = qy + r$ and $0 \leq |r| < |y|$.

Required Associated Types

type DivOutput

type ModOutput

Required Methods

fn div_mod(self, other: RHS) -> (Self::DivOutput, Self::ModOutput)