Trait malachite::num::arithmetic::traits::DivRem

pub trait DivRem<RHS = Self> {
    type DivOutput;
    type RemOutput;

    // Required method
    fn div_rem(self, other: RHS) -> (Self::DivOutput, Self::RemOutput);
}

Divides two numbers, returning the quotient and remainder. The quotient is rounded towards zero, and the remainder has the same sign as the dividend (first input).

The quotient and remainder satisfy $x = qy + r$ and $0 \leq |r| < |y|$.

Required Associated Types

type DivOutput

type RemOutput

Required Methods

fn div_rem(self, other: RHS) -> (Self::DivOutput, Self::RemOutput)