Trait malachite::num::arithmetic::traits::DivRound

pub trait DivRound<RHS = Self> {
    type Output;

    // Required method
    fn div_round(self, other: RHS, rm: RoundingMode) -> (Self::Output, Ordering);
}

Divides a number by another number and rounds according to a specified rounding mode. An Ordering is also returned, indicating whether the returned value is less than, equal to, or greater than the exact value.

Required Associated Types

type Output

Required Methods

fn div_round(self, other: RHS, rm: RoundingMode) -> (Self::Output, Ordering)