Trait malachite::num::arithmetic::traits::ModPowPrecomputed

pub trait ModPowPrecomputed<RHS = Self, M = Self>: Sized {
    type Output;
    type Data;

    // Required methods
    fn precompute_mod_pow_data(m: &M) -> Self::Data;
    fn mod_pow_precomputed(
        self,
        exp: RHS,
        m: M,
        data: &Self::Data
    ) -> Self::Output;
}

Raises a number to a power modulo another number $m$. The base must be already reduced modulo $m$.

If multiple modular exponentiations with the same modulus are necessary, it can be quicker to precompute some piece of data and reuse it in the exponentiation calls. This trait provides a function for precomputing the data and a function for using it during exponentiation.

Required Associated Types

type Output

type Data

Required Methods

fn precompute_mod_pow_data(m: &M) -> Self::Data

Precomputes some data to use for modular exponentiation.

fn mod_pow_precomputed(self, exp: RHS, m: M, data: &Self::Data) -> Self::Output

Object Safety

This trait is not object safe.