Trait malachite::num::arithmetic::traits::NegModAssign

pub trait NegModAssign<RHS = Self> {
    // Required method
    fn neg_mod_assign(&mut self, other: RHS);
}

Divides the negative of a number by another number, replacing the first number by the remainder.

If the quotient were computed, the quotient and remainder would satisfy $x = qy - r$ and $0 \leq r < y$.

Required Methods

fn neg_mod_assign(&mut self, other: RHS)