Trait malachite::num::arithmetic::traits::RoundToMultipleOfPowerOf2

pub trait RoundToMultipleOfPowerOf2<RHS> {
    type Output;

    // Required method
    fn round_to_multiple_of_power_of_2(
        self,
        pow: RHS,
        rm: RoundingMode
    ) -> (Self::Output, Ordering);
}

Rounds a number to a multiple of $2^k$, according to a specified rounding mode. An Ordering is also returned, indicating whether the returned value is less than, equal to, or greater than the original value.

Required Associated Types

type Output

Required Methods

fn round_to_multiple_of_power_of_2( self, pow: RHS, rm: RoundingMode ) -> (Self::Output, Ordering)