Trait malachite::num::arithmetic::traits::RoundToMultipleOfPowerOf2Assign

pub trait RoundToMultipleOfPowerOf2Assign<RHS> {
    // Required method
    fn round_to_multiple_of_power_of_2_assign(
        &mut self,
        pow: RHS,
        rm: RoundingMode
    ) -> Ordering;
}

Rounds a number to a multiple of $2^k$ in place, according to a specified rounding mode. An Ordering is returned, indicating whether the returned value is less than, equal to, or greater than the original value.

Required Methods

fn round_to_multiple_of_power_of_2_assign( &mut self, pow: RHS, rm: RoundingMode ) -> Ordering