Trait malachite::num::arithmetic::traits::SaturatingSubMul

pub trait SaturatingSubMul<Y = Self, Z = Self> {
    type Output;

    // Required method
    fn saturating_sub_mul(self, y: Y, z: Z) -> Self::Output;
}

Subtracts a number by the product of two other numbers, saturating at the numeric bounds instead of overflowing.

Required Associated Types

type Output

Required Methods

fn saturating_sub_mul(self, y: Y, z: Z) -> Self::Output