Trait malachite::num::arithmetic::traits::ShlRoundAssign

pub trait ShlRoundAssign<RHS> {
    // Required method
    fn shl_round_assign(&mut self, other: RHS, rm: RoundingMode) -> Ordering;
}

Left-shifts a number (multiplies it by a power of 2) in place, rounding the result according to a specified rounding mode. An Ordering is also returned, indicating whether the assigned value is less than, equal to, or greater than the exact value.

Rounding might only be necessary if other is negative.

Required Methods

fn shl_round_assign(&mut self, other: RHS, rm: RoundingMode) -> Ordering