Trait malachite::num::arithmetic::traits::ShrRoundAssign

pub trait ShrRoundAssign<RHS> {
    // Required method
    fn shr_round_assign(&mut self, other: RHS, rm: RoundingMode) -> Ordering;
}

Right-shifts a number (divides it by a power of 2) in place, rounding the result according to a specified rounding mode. An Ordering is also returned, indicating whether the assigned value is less than, equal to, or greater than the exact value.

Rounding might only be necessary if other is positive.

Required Methods

fn shr_round_assign(&mut self, other: RHS, rm: RoundingMode) -> Ordering