Trait malachite::num::comparison::traits::EqAbs

pub trait EqAbs<Rhs = Self>
where
    Rhs: ?Sized,
{
    // Required method
    fn eq_abs(&self, other: &Rhs) -> bool;

    // Provided method
    fn ne_abs(&self, other: &Rhs) -> bool { ... }
}

Determines equality between the absolute values of two numbers.

Required Methods

fn eq_abs(&self, other: &Rhs) -> bool

Compares the absolute values of two numbers for equality, taking both by reference.

Provided Methods

fn ne_abs(&self, other: &Rhs) -> bool

Compares the absolute values of two numbers for inequality, taking both by reference.

Worst-case complexity

Same as the time and additional memory complexity of eq_abs.