Module malachite::num::conversion

Traits for converting to and from numbers, converting to and from strings, and extracting digits.

Modules

digits

Traits for working with the digits of numbers.

from

Traits for converting between different number types. The traits are WrappingFrom, SaturatingFrom, OverflowingFrom, ConvertibleFrom, and RoundingFrom.

half

JoinHalves and SplitInHalf, traits for joining the bits of two numbers or for splitting a number in half.

is_integer

IsInteger, a trait for determining whether a value is an integer.

mantissa_and_exponent

Traits for converting numbers to and from mantissa and exponent representations.

slice

Traits for converting slices to numbers, slices to Vecs, or numbers to Vecs. The traits are FromOtherTypeSlice, VecFromOtherTypeSlice, and VecFromOtherType.

string

Traits for converting numbers to and from Strings.

traits

Various traits for converting numbers.