Trait malachite::num::conversion::traits::RoundingFrom

pub trait RoundingFrom<T>: Sized {
    // Required method
    fn rounding_from(value: T, rm: RoundingMode) -> (Self, Ordering);
}

Converts a value from one type to another, where the conversion is made according to a specified RoundingMode. An Ordering is also returned, indicating whether the returned value is less than, equal to, or greater than the original value.

Required Methods

fn rounding_from(value: T, rm: RoundingMode) -> (Self, Ordering)

Object Safety

This trait is not object safe.