Trait malachite::num::conversion::traits::SaturatingFrom

pub trait SaturatingFrom<T>: Sized {
    // Required method
    fn saturating_from(value: T) -> Self;
}

Converts a value from one type to another, where if the conversion is not exact the result is set to the maximum or minimum value of the result type, whichever is closer.

Required Methods

fn saturating_from(value: T) -> Self

Object Safety

This trait is not object safe.