Module malachite::num::exhaustive

Iterators that generate numbers without repetition.

Structs

ConstantPrecisionPrimitiveFloats

Generates all finite positive primitive floats with a specified sci_exponent and precision.

ExhaustiveNegativeFinitePrimitiveFloats

Generates all negative finite primitive floats.

ExhaustiveNonzeroFinitePrimitiveFloats

Generates all nonzero finite primitive floats.

ExhaustivePositiveFinitePrimitiveFloats

Generates all positive finite primitive floats.

ExhaustivePrimitiveFloatsWithExponent

Generates all positive finite primitive floats with a specified sci_exponent.

PrimitiveFloatIncreasingRange

Generates all primitive floats in an interval, in increasing order.

PrimitiveIntIncreasingRange

Generates all primitive integers in an interval.

Enums

ExhaustivePrimitiveFloatInclusiveRange

Generates all primitive floats in an interval.

ExhaustiveSignedRange

Generates all values of a signed integer type in an interval, in order of increasing absolute value.

Functions

exhaustive_finite_primitive_floats

Generates all finite primitive floats.

exhaustive_natural_signeds

Generates all natural (non-negative) signed integers in ascending order.

exhaustive_negative_finite_primitive_floats

Generates all negative finite primitive floats.

exhaustive_negative_primitive_floats

Generates all negative primitive floats.

exhaustive_negative_signeds

Generates all negative signed integers in descending order.

exhaustive_nonzero_finite_primitive_floats

Generates all nonzero finite primitive floats.

exhaustive_nonzero_primitive_floats

Generates all nonzero primitive floats.

exhaustive_nonzero_signeds

Generates all nonzero signed integers in order of increasing absolute value.

exhaustive_positive_finite_primitive_floats

Generates all positive finite primitive floats.

exhaustive_positive_primitive_floats

Generates all positive primitive floats.

exhaustive_positive_primitive_ints

Generates all positive primitive integers in ascending order.

exhaustive_primitive_float_inclusive_range

Generates all primitive floats in the closed interval $[a, b]$.

exhaustive_primitive_float_range

Generates all primitive floats in the half-open interval $[a, b)$.

exhaustive_primitive_floats

Generates all primitive floats.

exhaustive_primitive_floats_with_sci_exponent

Generates all positive finite primitive floats with a specified sci-exponent.

exhaustive_primitive_floats_with_sci_exponent_and_precision

Generates all finite positive primitive floats with a specified sci_exponent and precision.

exhaustive_signed_inclusive_range

Generates all signed integers in the closed interval $[a, b]$, in order of increasing absolute value.

exhaustive_signed_range

Generates all signed integers in the half-open interval $[a, b)$, in order of increasing absolute value.

exhaustive_signeds

Generates all signed integers in order of increasing absolute value.

exhaustive_unsigneds

Generates all unsigned integers in ascending order.

finite_primitive_floats_increasing

Generates all finite primitive floats, in ascending order.

negative_finite_primitive_floats_increasing

Generates all finite negative primitive floats, in ascending order.

negative_primitive_floats_increasing

Generates all negative primitive floats, in ascending order.

nonzero_finite_primitive_floats_increasing

Generates all finite nonzero primitive floats, in ascending order.

nonzero_primitive_floats_increasing

Generates all nonzero primitive floats, in ascending order.

positive_finite_primitive_floats_increasing

Generates all finite positive primitive floats, in ascending order.

positive_primitive_floats_increasing

Generates all positive primitive floats, in ascending order.

primitive_float_increasing_inclusive_range

Generates all primitive floats in the closed interval $[a, b]$, in ascending order.

primitive_float_increasing_range

Generates all primitive floats in the half-open interval $[a, b)$, in ascending order.

primitive_floats_increasing

Generates all primitive floats, except NaN, in ascending order.

primitive_int_increasing_inclusive_range

Generates all primitive integers in the closed interval $[a, b]$, in ascending order.

primitive_int_increasing_range

Generates all primitive integers in the half-open interval $[a, b)$, in ascending order.