Trait malachite::num::logic::traits::BitBlockAccess

pub trait BitBlockAccess: Sized {
    type Bits;

    // Required methods
    fn get_bits(&self, start: u64, end: u64) -> Self::Bits;
    fn assign_bits(&mut self, start: u64, end: u64, bits: &Self::Bits);

    // Provided method
    fn get_bits_owned(self, start: u64, end: u64) -> Self::Bits { ... }
}

Defines functions that access or modify blocks of adjacent bits in a number.

Required Associated Types

type Bits

Required Methods

fn get_bits(&self, start: u64, end: u64) -> Self::Bits

Extracts a block of bits whose first index is start and last index is end - 1.

fn assign_bits(&mut self, start: u64, end: u64, bits: &Self::Bits)

Assigns the least-significant end - start bits of bits to bits start through end - 1 (inclusive) of self.

Provided Methods

fn get_bits_owned(self, start: u64, end: u64) -> Self::Bits

Extracts a block of bits whose first index is start and last index is end - 1, taking ownership of self.

Worst-case complexity

For the default implementation, same as get_bits.

Panics

For the default implementation, ee panics for get_bits.

Examples

use malachite_base::num::logic::traits::BitBlockAccess;

assert_eq!((-0x5433i16).get_bits_owned(4, 8), 0xc);
assert_eq!((-0x5433i16).get_bits_owned(5, 9), 14);
assert_eq!((-0x5433i16).get_bits_owned(5, 5), 0);
assert_eq!((-0x5433i16).get_bits_owned(100, 104), 0xf);

Object Safety

This trait is not object safe.