Trait malachite::num::logic::traits::BitScan

pub trait BitScan {
    // Required methods
    fn index_of_next_false_bit(self, start: u64) -> Option<u64>;
    fn index_of_next_true_bit(self, start: u64) -> Option<u64>;
}

Defines functions that search for the next true or false bit in a number, starting at a specified index and searching in the more-significant direction.

Required Methods

fn index_of_next_false_bit(self, start: u64) -> Option<u64>

Given a number and a starting index, searches the number for the smallest index of a false bit that is greater than or equal to the starting index.

fn index_of_next_true_bit(self, start: u64) -> Option<u64>

Given a number and a starting index, searches the number for the smallest index of a true bit that is greater than or equal to the starting index.