Module malachite::sets::exhaustive

Iterators that generate sets without repetition.

Functions

exhaustive_b_tree_sets

Generates BTreeSets with elements from a single iterator.

exhaustive_b_tree_sets_fixed_length

Generates BTreeSets of a given size with elements from a single iterator.

exhaustive_b_tree_sets_length_inclusive_range

Generates BTreeSets, with lengths in a range $[a, b]$, with elements from a single iterator.

exhaustive_b_tree_sets_length_range

Generates BTreeSets, with lengths in a range $[a, b)$, with elements from a single iterator.

exhaustive_b_tree_sets_min_length

Generates BTreeSets with a mininum length, with elements from a single iterator.

exhaustive_hash_sets

Generates HashSets with elements from a single iterator.

exhaustive_hash_sets_fixed_length

Generates HashSets of a given size with elements from a single iterator.

exhaustive_hash_sets_length_inclusive_range

Generates HashSets, with lengths in a range $[a, b]$, with elements from a single iterator.

exhaustive_hash_sets_length_range

Generates HashSets, with lengths in a range $[a, b)$, with elements from a single iterator.

exhaustive_hash_sets_min_length

Generates HashSets with a mininum length, with elements from a single iterator.

lex_b_tree_sets

Generates BTreeSets with elements from a single iterator.

lex_b_tree_sets_fixed_length

Generates BTreeSets of a given size with elements from a single iterator.

lex_b_tree_sets_length_inclusive_range

Generates BTreeSets, with lengths in a range $[a, b]$, with elements from a single iterator.

lex_b_tree_sets_length_range

Generates BTreeSets, with lengths in a range $[a, b)$, with elements from a single iterator.

lex_b_tree_sets_min_length

Generates BTreeSets with a mininum length, with elements from a single iterator.

lex_hash_sets

Generates HashSets with elements from a single iterator.

lex_hash_sets_fixed_length

Generates HashSets of a given size with elements from a single iterator.

lex_hash_sets_length_inclusive_range

Generates HashSets, with lengths in a range $[a, b]$, with elements from a single iterator.

lex_hash_sets_length_range

Generates HashSets, with lengths in a range $[a, b)$, with elements from a single iterator.

lex_hash_sets_min_length

Generates HashSets with a mininum length, with elements from a single iterator.

shortlex_b_tree_sets

Generates BTreeSets with elements from a single iterator.

shortlex_b_tree_sets_length_inclusive_range

Generates BTreeSets, with lengths in a range $[a, b]$, with elements from a single iterator.

shortlex_b_tree_sets_length_range

Generates BTreeSets, with lengths in a range $[a, b)$, with elements from a single iterator.

shortlex_b_tree_sets_min_length

Generates BTreeSets with a mininum length, with elements from a single iterator.

shortlex_hash_sets

Generates HashSets with elements from a single iterator.

shortlex_hash_sets_length_inclusive_range

Generates HashSets, with lengths in a range $[a, b]$, with elements from a single iterator.

shortlex_hash_sets_length_range

Generates HashSets, with lengths in a range $[a, b)$, with elements from a single iterator.

shortlex_hash_sets_min_length

Generates HashSets with a mininum length, with elements from a single iterator.